--- ansible/hogwild.yml
@@ -1,47 +1,0 @@
----
-
-- hosts: localhost
-  gather_facts: yes
-  tasks:
-    - name: "Clone repo"
-      git:
-        accept_hostkey: yes
-        repo: "git@github.com:Shillaker/hogwild.git"
-        dest: "/usr/local/code/hogwild"
-        clone: yes
-        update: yes
-
-    - name: "Check for data"
-      stat:
-        path: "/usr/local/code/hogwild/data"
-      register: data
-
-    - name: "Download hogwild data"
-      shell: "bash download_data.sh"
-      args:
-        chdir: "/usr/local/code/hogwild"
-      when: data.stat.exists == false
-
-    - name: "Create build dir"
-      shell: "{{ item }}"
-      args:
-        chdir: "/usr/local/code/hogwild"
-      with_items:
-        - "rm -rf build"
-        - "mkdir build"
-
-    - name: "Build hogwild"
-      shell: "{{ item }}"
-      args:
-        chdir: "/usr/local/code/hogwild/build"
-      with_items:
-        - "cmake .."
-        - "make"
-        - "mv bin .."
-      environment:
-        CXX: "/usr/bin/clang++"
-        CC: "/usr/bin/clang"
-        CPP: "/usr/bin/clang-cpp"
-
-    # Hogwild command:
-    # bin/svm_generic --mu 0.01 --binary 1 --splits 5 --stepinitial 0.01 --step_decay 0.8 data/RCV.binary.train.dat data/RCV.test.dat
