+++ src/runner/CMakeLists.txt
@@ -40,9 +40,6 @@ target_link_libraries(bench_mem bench)
 add_executable(bench_func bench_func.cpp)
 target_link_libraries(bench_func bench)
 
-add_executable(bench_omp bench_omp.cpp)
-target_link_libraries(bench_omp bench)
-
 add_executable(bench_time bench_time.cpp)
 target_link_libraries(bench_time bench)
