+++ workon.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+#CALLING_DIR=$(pwd)
 THIS_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 pushd ${THIS_DIR} >> /dev/null
 
@@ -14,6 +15,13 @@ fi
 export VIRTUAL_ENV_DISABLE_PROMPT=1
 source venv/bin/activate
 
+# Allow overriding the invoke tasks file
+# INVOKE_ROOT=${INVOKE_ROOT:-$THIS_DIR}  
+# alias inv="inv -r ${INVOKE_TASKS}"
+
+# Ensure both the Faasm Python code is accessible, as well as the calling dir
+export PYTHONPATH="${PYTHONPATH}:${CALLING_DIR}:${THIS_DIR}/tasks"
+
 # ----------------------------
 # Invoke tab-completion (http://docs.pyinvoke.org/en/stable/invoke.html#shell-tab-completion)
 # ----------------------------
@@ -24,7 +32,7 @@ _complete_invoke() {
     COMPREPLY=( $(compgen -W "${candidates}" -- $2) )
 }
 
-complete -F _complete_invoke -o default invoke inv
+# complete -F _complete_invoke -o default invoke inv
 
 # ----------------------------
 # Environment vars
